Refactored respondent details into common
